pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "airflow-watcher"
version = "1.0.0"
description = "Airflow UI Plugin for monitoring DAG failures and SLA misses"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Ramanujam Solaimalai", email = "ramanujam.solaimalai@gmail.com"}
]
maintainers = [
    {name = "Ramanujam Solaimalai", email = "ramanujam.solaimalai@gmail.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Apache Airflow",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "apache-airflow>=2.7.0,<3.0",
    "flask>=2.0.0",
    "requests>=2.28.0",
    "slack-sdk>=3.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
pagerduty = [
    "requests>=2.28.0",
]
statsd = []
prometheus = [
    "prometheus-client>=0.15.0",
]
alerting = [
    "requests>=2.28.0",
    "slack-sdk>=3.19.0",
]
all = [
    "requests>=2.28.0",
    "slack-sdk>=3.19.0",
    "prometheus-client>=0.15.0",
]
standalone = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=1.4.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
airflow-watcher-api = "airflow_watcher.api.main:main"

[project.entry-points."airflow.plugins"]
airflow_watcher = "airflow_watcher.plugins.watcher_plugin:AirflowWatcherPlugin"

[project.urls]
"Homepage" = "https://github.com/ram07eng/airflow-watcher"
"Bug Tracker" = "https://github.com/ram07eng/airflow-watcher/issues"
"Documentation" = "https://github.com/ram07eng/airflow-watcher#readme"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=airflow_watcher --cov-report=term-missing"
markers = [
    "load: load and stress tests (deselect with '-m \"not load\"')",
    "integration: tests that require a live Airflow DB (not run in CI by default)",
]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
no_implicit_optional = true
warn_unused_configs = true
warn_return_any = true
warn_unreachable = true
disable_error_code = ["import-untyped"]
